+++ kubejs/server_scripts/base/recipes/emendatusenigmatica/shapeless.js
@@ -16,6 +16,16 @@ ServerEvents.recipes((event) => {
             inputs: 'emendatusenigmatica:sulfur_cluster_shard_block',
             output: '4x emendatusenigmatica:sulfur_cluster',
             id: `${id_prefix}from_block/sulfur`
+        },
+        {
+            inputs: 'emendatusenigmatica:raw_osmium_block',
+            output: '9x emendatusenigmatica:raw_osmium',
+            id: `${id_prefix}from_block/osmium`
+        },
+        {
+            inputs: 'emendatusenigmatica:raw_tin_block',
+            output: '9x emendatusenigmatica:raw_tin',
+            id: `${id_prefix}from_block/tin`
         }
     ];