+++ changelogs/CHANGELOG.md
@@ -7,11 +7,13 @@
 ### 🌟 Improvements
 
 -   Fiery Metal and Liquid Blaze Burners may now be used for PNC heat [\#467](https://github.com/EnigmaticaModpacks/Enigmatica9/pull/467) ([MuteTiefling](https://github.com/MuteTiefling))
+-   Added Conversion between Limestone from Quark and Create [\#479](https://github.com/EnigmaticaModpacks/Enigmatica9/pull/479) ([MuteTiefling](https://github.com/MuteTiefling))
 
 ### 🐛 Fixed Bugs
 
 -   Fix Some Assembly Required sawing recipes [\#460](https://github.com/EnigmaticaModpacks/Enigmatica9/pull/460) ([MuteTiefling](https://github.com/MuteTiefling))
 -   Reduce spawn range for Spirit spawners [\#460](https://github.com/EnigmaticaModpacks/Enigmatica9/pull/460) ([MuteTiefling](https://github.com/MuteTiefling))
+-   Osmium and Tin Raw Ore Blocks now craft down into Raw Ore properly [\#479](https://github.com/EnigmaticaModpacks/Enigmatica9/pull/479) ([MuteTiefling](https://github.com/MuteTiefling))
 
 ---